Refactored out the RoomContext
