+++ frontend/src/context/RoomActions.js
@@ -1,0 +1,18 @@
+const RoomActions = {
+    connectedToRoom: "CONNECTED_TO_ROOM",
+    connectionClosed: "CONNECTION_CLOSED",
+    setUserName: "SET_USER_NAME",
+    userNamesUpdated: "USER_NAMES_UPDATED",
+    itemAdded: "ITEM_ADDED",
+    itemDeleted: "ITEM_DELETED",
+    itemsUpdated: "ITEM_UPDATED",
+    itemSelected: "ITEM_SELECTED",
+    boundsUpdated: "BOUNDS_UPDATED",
+    roomNameUpdated: "ROOM_NAME_UPDATED",
+    roomDeleted: "ROOM_DELETED",
+    loading: "LOADING",
+    error: "ERROR",
+    clearEditorActionQueue: "CLEAR_EDITOR_ACTION_QUEUE",
+};
+
+export default RoomActions;
