+++ frontend/src/context/initialState.js
@@ -1,0 +1,15 @@
+const initialState = {
+    items: null,
+    templateId: null,
+    roomName: null,
+    bounds: null,
+    loading: true,
+    error: null,
+    socketConnection: null,
+    userName: null,
+    userNames: [],
+    editorActionQueue: [],
+    selectedItemID: null,
+};
+
+export default initialState;
